+++ cg/models/mip/mip_metrics_deliverables.py
@@ -1,6 +1,6 @@
-from typing import List, Optional, Any
+from typing import Any, List, Optional
 
-from pydantic import BaseModel, validator, Field
+from pydantic import BaseModel, Field, validator
 
 from cg.constants.gender import Gender
 
@@ -70,6 +70,14 @@ class MeanInsertSize(BaseModel):
         return int(value)
 
 
+class MedianTargetCoverage(BaseModel):
+    """Definition of median target coverage"""
+
+    sample_id: str
+    step: str
+    value: int
+
+
 class ParsedMetrics(BaseModel):
     """Defines parsed metrics"""
 
@@ -80,6 +88,8 @@ class ParsedMetrics(BaseModel):
     mapped_reads_step: str
     mean_insert_size: int
     mean_insert_size_step: str
+    median_target_coverage: int
+    median_target_coverage_step: str
     predicted_sex: str = Gender.UNKNOWN
     predicted_sex_step: str
 
@@ -93,6 +103,7 @@ class MetricsDeliverables(BaseModel):
     mapped_reads: Optional[List[MappedReads]]
     mean_insert_size: Optional[List[MeanInsertSize]]
     predicted_sex: Optional[List[GenderCheck]]
+    median_target_coverage: Optional[List[MedianTargetCoverage]]
     sample_id_metrics: Optional[List[ParsedMetrics]]
 
     @validator("sample_ids", always=True)
@@ -152,6 +163,18 @@ class MetricsDeliverables(BaseModel):
                 )
         return mean_insert_size
 
+    @validator("median_target_coverage", always=True)
+    def set_median_target_coverage(cls, _, values: dict) -> List[MedianTargetCoverage]:
+        """Set median target coverage"""
+        median_target_coverage: List = []
+        raw_metrics: List = values.get("metrics_")
+        for metric in raw_metrics:
+            if metric.name == "MEDIAN_TARGET_COVERAGE":
+                median_target_coverage.append(
+                    MedianTargetCoverage(sample_id=metric.id, step=metric.step, value=metric.value)
+                )
+        return median_target_coverage
+
     @validator("predicted_sex", always=True)
     def set_predicted_sex(cls, _, values: dict) -> List[GenderCheck]:
         """Set predicted sex"""
@@ -173,6 +196,7 @@ class MetricsDeliverables(BaseModel):
             "duplicate_reads": values.get("duplicate_reads"),
             "mapped_reads": values.get("mapped_reads"),
             "mean_insert_size": values.get("mean_insert_size"),
+            "median_target_coverage": values.get("median_target_coverage"),
             "predicted_sex": values.get("predicted_sex"),
         }
         for sample_id in sample_ids: